requirements.txt
blinker==1.6.2
casefy==0.1.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coco-cfp @ file:///root/coco_cfp-0.0.3-py3-none-any.whl
coco-cfp-pfm-es @ file:///root/coco_cfp_pfm_es-0.0.5-py3-none-any.whl
confluent-kafka==2.2.0
dacite==1.8.1
dataclasses-avroschema==0.47.2
dnspython==2.4.2
fastavro==1.8.2
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
Inflector==3.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
numpy==1.25.2
pandas==2.0.3
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
werkzeug==2.3.7
zipp==3.16.2
